NPC_Dialogue_System_MCP/requirements.txt
fastapi
uvicorn
langchain
langgraph
tiktoken
requests
pyyaml